lint: clear all clippy warnings + add clippy CI gate
- Remove dead code: normalize_text (dup of helpers), add/remove_banned_word,
  scan_banned_words, command_error + redundant normalize_text_basic test
- export.rs: collapse identical if/else branches
- commands/mod.rs: filter_map -> filter, drop unnecessary mut
- models.rs: document reserved format field (kept for wire contract)
- ci.yml: add cargo clippy --all-targets -D warnings to macOS+Windows matrix

clippy --all-targets -D warnings now passes. 40/40 tests pass.

+++ src-tauri/src/commands/schema.rs
@@ -403,7 +403,7 @@ mod tests {
             project_path: "/tmp/test.grimoire".to_string(),
             database_path: ":memory:".to_string(),
             app_version: "1.0.0".to_string(),
-            schema_version: SCHEMA_VERSION as i64,
+            schema_version: SCHEMA_VERSION,
             created_at: String::new(),
             updated_at: String::new(),
         };
